pyproject.toml
[tool.poetry]
name = "ebible"
version = "0.1.0"
description = ""
authors = ["Partnership for Applied Biblical NLP (PAB-NLP)"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
tqdm = "^4.64.1"
pandas = "^1.5.2"
requests = "^2.28.1"
lxml = "^4.9.1"
daiquiri = "^3.2.1"
beautifulsoup4 = "^4.11.1"
regex = "^2022.10.31"
isort = "^5.11.4"
PyYAML = "^6.0"
sil-machine = "1.3.0"
openpyxl = "^3.1.5"
python-dotenv = "^1.1.0"
pytest = "^8.3.5"
pyarrow = "^20.0.0"
huggingface-hub = "^0.30.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
